enhance: move jemalloc_stats from pkg to internal/util/segcore
Move jemalloc_stats.go and its test file from pkg/util/hardware to
internal/util/segcore. This is a more appropriate location because:
- jemalloc_stats depends on milvus_core C++ library via cgo
- The pkg directory should remain independent of internal C++ dependencies
- segcore is the natural home for core memory allocator utilities
